Resolved: #0022921: No tip when pics shown very slowly
Made the "needed pictures" estimate persist up to GetPostMakeStatus.
This gets us the right "slow pictures" tip in most cases.
Note that the behaviour when trying to stretch pictures by
more than a limit (8x currently), a normal paced muvee will be
made and the music will be clipped. Therefore there is no case
such that pictures will be slowed down *and* music will be clipped.