#0023215: Need style (or feature) to view music descriptors
First drop of visual music descriptors code.
A new effect, MusicDescriptors has been added to create simple graphics for cut,flash and loudness hints. Cut hints and Flash hints are represented by an animated circle. Loudness hints are represented by a simple vertical meter.

stylelib2.scm has been modified to include the difinitions and instantiations of each hint effect.

Textout effect has been modified so that the rendered text can be translated about on the screen. This is very useful in the musicDescriptor style as we want labels for each hint. Previously, this could not be done.


That's how the code should be used:

(define muvee-global-effect
   (effect-stack
      (effect "Perspective" (A))
       visual-loudness-hints
       visual-flash-hints
       visual-cut-hints
      ;visual-flash-hints-values    
      ;visual-cut-hints-values	  	 
      ;visual-loudness-hints-values 
))