Resolved: #0023303: Application crashes when making muvee with split videos and music
For the case where the expected muvee duration is more than can be
done using input videos (without pictures), this condition is triggered.
This is because the "try to expand segments" phase of construction does
a comparison against a "normal" construction to keep changes predictable.
This comparison was failing since the introduction of code to split/merge 
segments around music boundaries and when shot boundaries cannot be 
straddled due to usage constraints, which happens because segments are split
in the UI, indicating that the user would like to have something from each
of the split clips if possible.