#0023215: Need style (or feature) to view music descriptors
The duration of a hint, the persistance of the visual graphic and whether it is a cicle of a rectangular bar can now be specified from the style code. stylelib2.scm has been modified accordingly to receive those values. The number of hints per minute has also been increased to the manximum (1199) so that all hints are included. Unused code in stylelib2.scm has been deleted and the remaining code has been cleaned  up.