Resolved: #0023387: map function doesn't behave correctly with lazy lists
The list that is being worked on by the list_map
function wasn't protected from a garbage collection
that might get triggered from muse_tail that was being
called *after* the _unwind(sp). The muse_tail() needed
to be called *before* the _unwind(sp) for GC to work
correctly in this context and keep the working list
protected.