0023306: Previewing of muvee do not contain video sondtrack even though had manually set video soundtrack to be audible in muvee playback
The m_renderInfo struct doesn't contain valid data any more.
The only field that is valid is the aspect ratio. All others
are set to constant. i.e. the constructor can no longer 
construct a timeline according to the output frame size or
frame rate, but only according to the output aspect ratio.
Therefore this commit obsoletes all other fields by explicitly
removing the structure and replacing it with the only item
that is valid. So m_renderInfo is gone and you'll find 
m_renderAspectRatio in its place.

MiscLingo was exposing all the fields of m_renderInfo to 
the muSE code through render-frame-width, render-frame-rate,
etc. Now they are all gone, except render-aspect-ratio.
I checked that only render-aspect-ratio is used in stylelib2.scm
and all the data.scm files of current styles, therefore this is
a safe change.