RESOLVED: #0022797: Corrupt thumbnail in muveeShare dialog
Inside FImageTexture, there is a function called Update_glTexSubImage2D_Perf() that is called when muvee reveal is first launched. What that function does is draw a texture and measure the time it takes to do so. That result is used to determine whether to enable AMM or not. 

The issue here is that after performing the test, we were not clearing the screen. So when we made a muvee, we ended up with artefacts on the screen. These artefacts were only present in certain styles such as Ultra Plain and Classic Vanilla. The reason being that the title section of those two styles have transparent backgrounds with text in front. So we ended up seeing what had been previously drawn under the transparent region.

Anyways now after doing the performance test, we actually clear the screen. That fixes the problem.