resolved:#0022924
Fix the location of tooltip for control buttons so it always appears within the panel. 

This seems to be Windows Refresh problem and happens only on that minipc. So after discussed with #yeecheng, only check in for this branch for now, but do take note if it happens on other pc again.