#0017742: Need Lighting primitive for Journal
We were setting the material properties inside EvalEffect(). That's not good because if we have a delayedDrawing with an ImageOp, we can actually skip EvalEffect() and go straight to ImageOp().

Now that's really bad because we'll have an imageOp where none of the material properties are set. This is especially true for varying parameters like Alpha!

That has now been fixed by moving the grid setup inside ImageOp and well as the AlphaUpdate.