0023728: [Summer09] (hp# 87373) FOXCONN:Pop up a unlocalized message about muvee after change the color to 16bits.
Checked in localized strings for the popup message - English, French and Japanese.
mES now has a SetLocale() function that can be called to set the locale in which 
such error popups have to be displayed.

TODO: Ernest to call mES's SetLocale() whenever the app updates the locale of the
MVCOre environment object.