#0015556: Rolling credit
I went through all the muvee Reveal shipping styles and 4 of them had issues with the rolling credits. They are:

1. MixMatch1 CorkBoard
2. MixMatch2 LoveStruck
3. MixMatch3 Industrial
4. Reflection2 Reflections.

These issues were all solved by setting the title and credits alpha to 0.99 and additionally, in the case of the MixMatch pack, the title vertical motion has been restricted to downwards. If the title moves up, you end up with the text and first user picture moving at the same time and both being visible on the screen. It's just messy.