Resolved: #0023530: video soundtrack using the sample video loops duing muvee preview.
1)audio loop back
2)audio not sync with video
Now we are using FlattenTimeMap to map the the playtime to mediaTime in 512 boundry interval
If there is left over media sample in any call it will be appended in next call to avoid 
audio disruption. Which address the both problem as given above

3) when output is mono and input is stereo during saving Reveal crash
4) when input is mono and output is stereo there is noise in one channel 
Input sound buffer will be initialized to maximum of input file channel 
and sampleRateConverter will be initialized to max of inputFile channle
or PlayChannel to handle stereo and mono properly. This address the above two problems