#0015556: Rolling credit
Textout.cpp

I had forgotten to multiply the heightFactor to the heightwidth and baseHeight of the 16:9 output. This has been corrected.

TextBase.cpp

The height of a character is now properly calculated in pixels. This will allow us to resize our output layer according to the number of lines used in Rolling credits.