0023146: WTV input support
WTV change the frame rate information after playing the File
code changes will play the graph for 500 ms to get the actual 
frame rate information.

While releasing the Graph PBDA DTFilter hangs we actually need
to call the release on PBDA Filter twice to get it work this is
hack for a while untill we know the reason why PBDA Filter hangs 
on first time.