Resolved #0023429: DirectShowVideo.cpp(627) ASSERT FAILED:
We can get these assert with minor timing discrepancy. Assert mention in the bug note,
as we are actually seeking to end of the file where as media is already at the end of the File
this case is handled by setting the seek request to start of the file.