#0015556: Rolling credit
HeightFactor has been coverted to type double. That should and does remove the lag between the start of the rolling and the actual visualization of the rolling credits. One other advantage of this is that the rolling speed now increases linearly as the number of lines increases instead of increases in steps.