0024049: Add support for multiple inputs in styles 0013043: muSE-X related changes
#tengchwan faced a crash when trying to use the show-duration-secs
and such functions with an invalid argument. The functions should
put up proper error messages and not crash. This commit adds
checks to GetShow() which is used by these functions and will put
up an appropriate error message when given invalid arguments.

.. another change is to muse_similar_symbol to give more relevant
similar symbols in error messages. For example, you can now use
it to get a similar symbol that is defined.