0024048: win32 service running under Vista is unable to acquire a rendering context .
Check in codes to demo the problem of getting "GDI Generic" RC from Vista Service.

Two build configurations [Release_TestServiceCreateRC] and [Debug_TestServiceCreateRC] are created for this purpose.

=== How it works ===
The console generates "c:\MVTestRenderer.txt", which tells the renderer string and pixel format for the RC created.

Note that on Vista, running the console from cmd creates the expected renderer but "GDI Generic" is created if we start from the service. One of the reason is that "ChoosePixelFormat()" returns different pixel formats for the two cases.