#Resolved #0023912: Switching to FSP does not resume muvee playback for video source with captions applied
You do not really need to go to Full-screen preview to see the bug. All you have to do is press the Pause button and then click on Play again.

Ok here's what was happening. Inside muveeReveal.sln, everytime the play button is triggered, we check if the current settings for the muvee has changed. This is done by creating a temporary project called dirtyCompareProj.rvl and diffing it with lastmade.rvl which, as its name suggests, has the settings of the last made muvee. If the two files are different, we make muvee again, else we just play the muvee.

When the SDKRuntime project file was compiled in release mode, the bug was present, but in debug mode, the bug was not present. That suggested that something was not being initialized properly. After adding printf all over the place, I found out that the LogFont structure that is used to store captions was not properly initialized in GetCaptionInfoCount(..) found in MediaSource.cpp, which ironically is in CoreMVModules project. 

Essentially we did not initialize the ClipPrecision of the font and thus it ended up containing garbage at runtime. It's been like that since svn rev 3.