#0024087: Locate missing file: crashes when missing file is using back the same folder path
Fixed the same path causes crash bug. It is because that Reveal searches "\\" for folders of org and new path, and compare them to get the same part. In case where the path are exact same, the search for "\\" will return -1 eventually, which causes the exception when retrieve folder name.