0024048: win32 service running under Vista is unable to acquire a rendering context .
Implement a simplified version of Vista service hack, mentioned in:
http://www.codeproject.com/KB/vista-security/VistaSessions.aspx

The main idea is to create process with a duplicated user token of "winlogon.exe",
in order to get maximum access rights (to desktop), for process created in services on Vista.
In this way, we get hardware accelerated render context.