0016490: Constructor bugs - The Collectors Edition
Removed dead code in the constructor. Now the inflating
constructor is always the one to be used, we can dicard
the pre-R9 constructor code.

Also, the older constructor did a double pass on ComposeMedia().
This was essential for the older algo, but is not needed for
the inflating constructor. So I've removed the second pass.