#0015556: Rolling credit
Merged changes in rev 9512 to BR3_Branch. This prevents downward push transition for both Title and Credits, which look messy when used with Rolling text feature. This will be disabled in the next build of MixMatch stylePack.

Note: Reflections2 and Showreel styles have issues with Rolling text, but these two styles are bundled together with the app. The Rolling text feature, together with a new build of the default styles, will be built from the trunk.