0024284: Weird pacing behaviour with split clips.
1. Removed older segment splitting code. This code is not 
needed with recent round of changes to highlight handling 
and is interfering with regular allocation.

2. Added segment splitting at the point where the constructor
realizes that entire video clips will need to be dropped if segments
are not split.

TESTS: 1 and 2 work in this case, video-only, slow and fast pacing. 
Haven't tested with highlights yet.

UNFINISHED: "N video clips were dropped" tips is not shown when 
clips are dropped in their entirety.

PROBLEMS: 

1. If duration is so short that only pictures can be fit,
then long picture segments get shown and all video gets dropped.
This is a discontinuity in behaviour as you gradually reduce
duration. Probably easily fixed.

2. If pictures are squashed to accommodate video clips, then
we don't have an appropriate tip to say "remove xx clips". In this
case, the clips can be either photos or video.