0024298: XMLUtils2 when loading a non encrypted xml file with white space at the beginning of the document - fails.
The cause was that the FileIsPlainXML function detected the file as encrypted due to the white spaces. Therefore the test is now done simply by loading the file directly first - if that succeeds we assume that its not encrypted, if it doesnt, we decrypt the contents and try and load it. If that doesn't work either - its *not a valid xml file*.