0024306: A highlight that was split into two using "split into clips" or manual splitting maybe skipped.
The highlight filter in the constructor was excluding all 
highlights that don't fall within clip bounds in their entirety.
As a consequence, highlights that end up being split due
to a "split into clips" operation or manual splitting
can end up being ignored.

This commit fixes the highlight filter to clip partially
overlapping highlights to the clip bounds. Note that 
the SDKRuntime is also expected to cleanup this data before
passing to the constructor. Broadly, the constructor has
to be given data that is independent of how it was arrived 
at using the UI.