Resolved: #0024087: Locate missing file: crashes when missing file is using back the same folder path
As suggested by Em:
I agree that we do not need to sync the thumbnail to the status of the file. However it would be good to replace the thumbnail with 'File not found' thumbnail as soon as we detect it. i.e. when the user previews or make a muvee. This way the user will know which files are missing.