Resolved #0024325: ASSERT FAILED: (t > m_segments.back().endt)
There is a minor problem in the initialization of parameter value (i.e. "CharHeight" for zooming in and out, "FadeText" for fading in and out, and "TranslateY" for rolling text.) By initializing it as:

  anim->Linear( t0, startVal );

It adds a new entry, t0=0, and thus asserting that this value should be bigger than zero, which is equal to this value. In other words, this entry is redundant. Instead, it should be initialized like this:

  anim->At( t0, startVal );

This sets the initial value to startVal and skips assert checking if this is the initial value. In fact, the Typewriter text animation is initialized this way and has no such aserts.