#0024287: Need background effects for pichu style.
LayerScaleX and LayerScaleY have been merged together.
The layer scaling is now done at the coordinates level instead of calls to glScaled(..) This allows us to dictate which layer we want to scale.
Fragment program 6 has been added.
Slight change in the modes. This will be expanded further in the next commit.