#0024287: Need background effects for pichu style.
Inside Mask effect, we can rotate, scale and translate the mask in the Texture space.
However those 3 operations are counter-intuitive to what we're used to within the model and perspective views.
To make it easier for the style author to use those 3 operations, their behaviours have been inverted to resemble that of the model and perspective views.

Additionally after we scale our texture, we center the image to make the final texture position more aesthetically pleasing.