0023180: Long highlights are not preserved as is. They are split up according to the style.
First cut at tackling the broken highlights problem.
Includes a few tweaks to the video allocation algorithm,
particularly the way transitions are treated. Prior to
merging segments for long highlights, the partitioning
is now done based on segment durations without taking
transitions into account. This seems to work well in
general. Subsequently, the allocation uses the transition
durations as well. If that isn't done, then we end up
with a situation where we've merged segments to which
a long highlight has been allocated, but the merged 
segment isn't long enough to accommodate the entire
highlight.