0013043: muSE-X related changes
Further to the previous changes (~71190, r10514), 
I've modified the binding generator to support
the (input k X) expressions as well. These must 
appear *before* the keyword arguments. For example -

(layers (A)
	(Translate () 
		(input 0 A)
		x: 0.5
		z: -0.5))

is ok, but the following is not ok -

(layers (A)
	(Translate () 
		x: 0.5
		(input 0 A)
		z: -0.5))

because it is equivalent to 

(layers (A)
	(effect "Translate" ()
		(param "x" 0.5 (input 0 A))
		(param "z" -0.5)))

which doesn't make sense.