Resolved: #23956: Visual list item's pop-up menu will close when the mouse leaves the pop-up menu or leaves the list item. When mouse capture is explicitly set to null/other control, the mouse-leave event gets triggered and this closes the pop-up menu. In the case where pop-up is loading, there is an attempt to set focus to the first menu item, this changes the mouse capture, which triggered the pop-up to close instead.
Hence when you select a new list item, >1 click is required to open the menu, and sometimes you will notice a drop-down arrow appearing briefly on the another item which is not selected. The fix here is to set focus to the first menu item only when we are navigating the pop-up menu using the keyboard.