0024562: Better cut point selection algorithm needed
1. Tweaked several aspects of the cut point selection and
segment scaling so that the segment split/join operations
kick in on fewer occasions, the time range for scanning for
cut point is smaller, the segment duration pattern is more
predictable, etc. This results in overall predictability
of the duration pattern when the music response is minimum.
High music response requirement still adds some warts.

2. Prototype (unused) code for LoudnessPacing - exploring
a technique to fit segments *and* take care of loudness
pacing as well as possible.