Resolved #24630: [muFAT] ShowTime_VideoFile.cpp(175) ASSERT FAILED: (mt.size() > 1)
The root cause of this failure is a bug that was hidden in the constructor's
fringe case handling. My constructor changes for cut point selection (#24562)
exposed this fringe case bug.

The bug being the constructor tries to make a faster muvee when it finds that
the amount of given video is insufficient to make a muvee of the given duration.
(Making a faster muvee stands some chance of success because transitions have
a reduced duration in faster productions). That logic is overall right,
but there was a bug in calculating the total available video duration, which
resulted in much smaller reported durations than were actually available.
The related line in 1863 of MediaPreallocator.cpp which used to call
cmap.contribution() instead of (the correct) cmap.shotContribution().

Fixed that and added a slew of asserts for very small segments so we can
catch this problem much early on.

Tests passed: 0022407.run, constructor-allstyles.run