#0024590: Need for Cg test Style.
Inverting the order of profiles  chosen.
We first try to choose FP30, then FP20 then FP1.

The previous order was FP1, FP20 and FP30.
The issue with the original order is that FP1 is the profile that is the most likely to pass. 
So we should try to get the latest (and better) profiles first. i.e FP30.

Do note that FP30 and FP20 are Nvidia specific profiles. That's not a big issue cos other cards still have FP1 to fall back on.