#0024590: Need for Cg test Style.
The is the first commit for the CgFX changes. As a recap, CgFX will allow us to do multipass rendering.
Prototype of CompileAndCache has been modified to accept parameters "StringOrFile" and "CgFX".

Setting the first parameter to 1 indicates that we are sending in a text file for compilation.
Setting the second parameter to 1 indicates that we are sending in a CgFX file for compilation.

The Cg compiler has been modified accordingly to accept and compile CgFX files.
Also I've added profile FP40 to the list of profiles we can choose from.

The CgFragmentProgram effect now has a seperate branch to handle CgFX rendering. More work needs to be done on that side. It looks a bit dodgy. Stay tuned!