0013043: muSE-X related changes
FEATURE: If a layer expression turns out to be (), then
the layers primitive will ignore it. Currently it'll pop
up a dialog box declaring it to be an error. This is
convenient to declare particular layers conditionally.

COMPATIBILITY: This is a new feature. So you have to pay
attention to compatibility issues. As long as a style
that uses this feature is only run with this version
(or later) of the constructor, it is fine.

*DO NOT USE* this feature in your style if it is
intended to run with earlier versions of the constructor.