0024590: Need for Cg test Style. 0022867: Need for Cg Compiler 0024574: Picchu effects 0013043: muSE-X related changes
Implemented support for dynamic animatable parameters in CgFragmentProgram
effect. Also modified the CGTestStyle to use the dynamic parameters.

Tests: Ran the CGTestStyle style to see if the parameters work as intended.
The dynamic parameters code only kicks in when the effect doesn't pre-declare
its parameters and therefore doesn't affect effects that don't use it.