#kumar #0024590: Need for Cg test Style.
We now have full support for GlobalMode in Cg. Essentially a Cg program has the option of not being an imageOp. Previously the globalMode was working but it did not have any parameter support for the Cg effect. Turns out that we do not need to send in the texture id if the Cg program is only dealing with one texture. So I have removed all references of texture parameters from the cg C++ code. The muse code remains unchanged.

The CgFx code has been modified accordingly.

#0024574: Picchu effects 
CgMask now has support for dynamic parameters.