0013043: muSE-X related changes
Showed some love to the SRender effect, using which you can stack what would otherwise have to be Image-Ops. The SRender effect had existed since the beginning of the universe (at least that of mES) but unused due to the performance penalty it imposes. There are some special cases where it *can* be used with lower performance penalty if the viewport size is reduced. So I've brushed it up in order to make it usable.

Example: You can now stack TextureSubset and Grayscale like this -

; Select "circle crop" mode of TextureSubset
(define circle-crop (fx:TextureSubset (A) mode: 1 radius: 0.5 x0: 0.5 y0: 0.5))

; Select grey scale mode of the Greyscale effect
(define grey  (fx:Greyscale (A) GlobalMode: 1))

; Insert a rendering operation.
(define render (fx:Render (A)))

(define muvee-segment-effect (effect-stack
                              circle-crop
                              render
                              grey))

The "render" step that sits between "grey" and "circle-crop" will ensure that circle-crop gets a texture to operate on by first rendering the grey effect on the scene and copying the result of the render to a texture.

The "Render" effect has two parameters you can use to control the quality/performance trade-off - "Width" and "Height", which dictate the dimensions of the viewport to use during the rendering. If the current viewport is larger than these dimensions, the smaller dimensions are used. If the current viewport is smaller, then the current viewport dimensions are used.