Resolved #24634: Assert failures in Machu style with background enabled
Problem was with the FunkyBackground effect which had some resource
preparation errors in the code. The assert failure was an indicator
of the resource preparation error. I've made the necessary fixes
and also placed elaborate comments at the relevant code sites
explaining why a certain step is done in a certain way.