Resolved #0024664: Saving to full AVC-HD format hangs at 99%
IF Saving progress reached 100% but saving graph haven;t generated the EC_COMPLETE message we will query the source filter if they have rendered all the frames. If all the frames are rendered we will stop the saving graph without waiting for saving graph to generate the EC_COMPLETE message.