Merge commit for  #0024717
a) added a post build step to copy the mvboostrapper executable into the package folder so that when a given installer is 'extracted' it can be runnable on its own
b) added a typedef for the string map provided by StaticResourceUnicode
c) re-ordered header for mvBootStrapperExtractDlg
d) added a status enum for use in the OnTimer method of mvBootStrapperExtractDlg
e) removed unused SetInstance method.
f) added a IsBootStrapperStuffed() method to evaluate whether the bootstrapper has been injected with a cab file.
g) depending on whether the bootstrapper is stuffed or not, the datapath will be overridden to point to the current working directory (if it hasn't been overridden already).
h) the contents of the default cab resource is now a guid which will be tested against to check if there is no cab file stuffed into the bootstrapper.