0019759: muFAT crashes and assert failures - The Collection
A bug in VirtualMusicDescriptors class' NearestHint method
was exposed by the addition of VirtualMusicDescriptors::FindStrongestHint.
Fixed it by ensuring that NearestHint will strictly adhere to
its protocol of never returning the same hint time again
even if SearchBefore or SearchAfter was requested.

This bug could've influenced the code in DescriptorLingo's
GetHintsBetween function which is called via muSE by styles.