#0024574: Picchu effects
CgMask has been modified so that the program is compiled not at the start of the muvee but when the effect is instantiated. The reason we do that is to ensure that there is no lag on low-end machines where the Cg compiler might take an unreasonably long time to perform the compilation.