#0022867: Need for Cg Compiler
Disabling the FP20 profile. For some reason, trivial Cg programs do not seem to compile with this profile. I have used the RGBtoYUV Cg program on F40, FP30, FP20 and arbfp1. All work fine expect FP20. It seems like variables need to be 'output modifiers' for their range to be able to exceed [-1.0, 1.0] in that specific profile. I'll play it safe and disable FP20.