0024562: Better cut point selection algorithm needed
Modifications to provide better tips. Not very confident of
this approach since I have tested only a fraction of the
code branches, but seems to work right even if a style uses
map-tc-subdomain.

The core modification is to preserve the natural warp duration
that is calculated before the normalization of the loudness
pacing integral. Using that as the time comparison point gives
correct estimates of too many or too few pictures.