0024562: Better cut point selection algorithm needed
Fixed dispartities in the application of segment duration
scaling factor transfer curve by normalizing the curve
over the desired duration of the muvee. This gives consistent
tips in a broad variety of styles (across the ages).

Tested with strips, classic vanilla and ultra plain which
span three different categories of segment pattern and loudness
response specification.