0017958: S00514_UltraPlain "Ultra Plain"
Modified UltraPlain so that it actually does cuts
when requested. Currently, it is doing a pseudo
transition that looks like a cut, thereby losing
some material from the result of a "use all photos
and video" production.

The trick is to define the symbol muvee-transition
to itself, which effectively marks it as "undefined".
In muSE, all undefined symbols have themselves as
their values.