0020482: Setting face bias to maximum in fit to music muvee results in an odd muvee duration and wrong tip.
Fixed a bug in the score computation loop where the start
index used for the contribution calculation is correct
only for the cases of video files without shot boundaries,
but is wrong for video files with shot boundaries that
are not split in the UI.

(CASE: 1min muvee in Classic Vanilla with face selector set to max,
input video duration = 7m50s)

1. With this bug fix and forcing the face descriptors to
alternate between 1.0 and 0.0 as is expected of the ideal
face detector, you get perfect face selection.

2. You get perfect selection even with 0.8 and 0.2 as the values.

3. With 0.7 and 0.3, you get 2 missed out of 19 segments (10% error).

4. With 0.6 and 0.4, you get lots of errors (about 50%).

(2,3,4 are artificial values which do not use the normalization that is
done to face probability estimates in the analyzer. All three of them
will normalize to 1.0/0.0 if the normalization is applied.)

RESULT: With the above artificial values used, I've shown that given adequate
separation between the face probabilities, the algorithm now does
select the faces correctly.

However, if I use the ORIGINAL face descriptor values, about 30% of
the segments are non-face (6/19) even when I make a 1min
muvee out of the test case video which is nearly 8mins long.