0020482: Setting face bias to maximum in fit to music muvee results in an odd muvee duration and wrong tip.
Tweaked the face bias handling function so that when bias > 0.0,
values lower than 0.35 are severely discounted and when bias < 0.0,
values > 0.65 are severely discounted. This gives better selection
behaviour with the artificial descriptor choices of even 0.6/0.4
in which case you get only 30% errors, and you get perfect selection
for even 0.7/0.3 as opposed to 10% error with the previous function.

With this function, the errors in dealing with the current
face probability values has decreased, though at the expense of
selecting multiple segments from the same shot due to
the specified face preference. There are several false negatives
in the analyzer's result, so this behaviour is acceptable
as long as the function works as expected for correct artificial
vaules of the descriptor.