0024562: Better cut point selection algorithm needed
Trimmed the end point of the muvee to the specified duration
using the pacing algorithm in a specific end case. Generally,
the constructed muvee will be accurate to about 1/2 of a segment's
duration and then segment splitting will be applied to trim
it to the given muvee, but with the absence of segment splitting,
the last segment can overshoot the duration. This change
addresses that problem.

Gerry brought up this problem with the Kinetic style, which
I wrongly thought was due to body->credits transition. With
sample media and music, the Kinetic style shows the "summarization
works better with video at least 2x muvee duration" tip.
The used video is indeed about 68 secs, so that condition
passes, but the muvee duration is slightly off from the
music duration which triggers that "for the sanity of
the user" condition as well.