0016490: Constructor bugs - The Collectors Edition
Tweaked the video speed control in the case where squashing/stretching
is being applied to the segment durations. In those cases, the video
speed warp will be adjusted as well, (almost) as though the pacing
slider were being moved about.

The effect of this change will be particularly visible in #24625,
if you enable the behaviour using c:\muveeDebug\ShowEachVideoClip.txt.
In that case, a lot of squash could be applied to show all video clips
and if you don't change the speed accordingly, the result looks
incoherent.