0013043: muSE-X related changes
Fixed fetch-uri to work correctly within Reveal. The problem
was that InternetCloseHandle *must* be called *before*
CommitUrlCacheEntry. Prior to the timeout/process related changes,
this was the case and therefore the function worked in both
Reveal and the command-line. After the timeout/process related
changes, this function call order changed. The altered call order
works only in the REPL and not within Reveal, where I didn't
initially test it.