0019759: muFAT crashes and assert failures - The Collection
The destroy function for the standard i/o ports wasn't set
and therefore the port data structures weren't being released.
This resulted in a memory leak that grew with the number of
muSE instantiations - i.e. make muvees.

Thanks to #ernest for catching this and to #ali for
narrowing down the case.