0022750: constructor hangs with specified media 0024562: Better cut point selection algorithm needed
Further to r11047, one code path of the constructor dealing with
not enough video condition has been rendered obsolete by the new
LoudnessPacing algorithm and this codepath is causing problems
now. Removed this code path.

Also changed the "abort on 100 iterations" condition to
"make the 10th iteration the final one" .. with a successful
construction instead of a failure.