#0024738: user defined Logo superimposed on the muvee
Added logo support in constructor for anderson.

Since we do not have UI support for it yet, it can only be activated with the presence of the file 'logo.txt' inside c:\muveeDebug

Inside that file, you need to set the file path of the logo as well as the positioning coordinates X1,X2,Y1,Y2.

Example of a logo.txt file

C:\\muveeDebug\\logo\\logo.png
0.0
0.1
0.8
0.9

The above example sets the logo to the top left of the screen.