#0024738: user defined Logo superimposed on the muvee
I added MVASSERTS to check for the range of X1,X2,Y1 and Y2.
I have made coordinates (0,0) to be the top left section of the screen. (this is the coordinate system the UI uses)
Fixed bug in the y-axis. It was previously going from 0 to 1. OpenGL however requires it to go from -1 to 1. 
Do note that the UI only deals with ranges (0,1) only. I do the necessary conversion programatically.