#0024928: Add "Film Color" option to all built-in styles
Renamed Greyscale as  Monochrome. Vivid Colors as Saturated.
Commented out HARDMIX blend mode. It's ugly to look at.
Re-enabled captions for Ultra-plain. Captions are unaffected by blend mode.
Removed debugging code.

Finally the main focus of this commit:

All the blend modes can now execute within 64 assembly instructions. We have that hard constraint because reveal needs to work on intel GMA machines. So a fair bit of optimization was done to the blend modes that had conditional branching. Essentially we perform the repetitive math outside the branch and re-use it within the different paths of the branches.