RESOLVED: #0024964: Rolling credit : Speed of rolling text maybe too fast
We now dynamically change the duration of titles and credits when the text animation is set to 'Rolling text'. 
All muvee styles have a title duration of 5s and a credits duration of 7s.

So what we are doing as from now is 

1. Calculate how many lines the text takes.
2. Get the font size
3. Get the total number of actual space of the screen that text will take.

Armed with this, we do the following:

1. If the text takes the space of about one text layer, we go for the usual 7s for credits.
2. However, if the text takes more that one one text layer (i.e we got lots of text rolling), we multiply the 'HeightFactor' (that's the variable used to expand the text layer vertically) with the credits duration.

eg: our text takes 1.5 text layers, credits duration will be 1.5 * 7s = 10.5s