#0006531: Music trimming [Constructor Changes]
We should only consider minMusicRepeatDuration_secs for the last song in the input music list.

This commit takes care of some edge case in which the last music in the input list is very short (5sec~7secs, possibly trimmed), observed by YeeCheng. The very short music would be ignored before this commit, due to the previous logic with minMusicRepeatDuration (5 secs). Now, we only consider minMusicRepeatDuration_secs for the last song in the input music list so that it could still be heard.