#0024928: Add "Film Color" option to all built-in styles
Changing default color tint fron 0xECEA6B (bright yellow) to 0xAFAD50 (dark yellow). 
We need to do this because we want a default tint that looks nice with most of the blend modes.
Currently we've got 17 blend modes that are very different from one another in terms of pixel manipulation.
That pretty much concludes the work on the blend modes. 

So far the blend modes have only been added to Ultra Plain. I'll now replicate the code to all default Anderson styles.