0023180: Long highlights are not preserved as is. They are split up according to the style.
Found (and fixed) a bug in the AllocateMedia function of MediaPreallocator where it
wouldn't correctly adjust the segments in the presence of long
highlights that span merged segments. This resulted in gaps in the
final production that can cause rendering to crash similar to #24966.

Although such gaps are not observed in the logs of case #24966, this fix
could contribute towards fixing it.