0024966: Crash during preview and save with some kinds of muvees
Put an additional guard in ShowTime_VideoClip to prevent cases
like this one. One possibility that occurred to me is that a
frame in requested of a VideoClip object at a time that is not
within its play interval. This could cause a crash in mES with
the characteristics reported in the log for these cases - i.e.
the mtIx2 assert failure.