#0024738: user defined Logo superimposed on the muvee [Top-level ticket]
MVCore was sending the wrong data to the constructor. I was getting X1 Y1 X2 Y1 as input. I've corrected it to X1 Y1 X2 Y2.

Ok the good news is we have a watermark feature that actually displays something on the screen. So end-to-end, it is working.

There seems to be an issue as to how the watermark is being resized. I'll talk to the the relevant parties.