#0024928: Add "Film Color" option to all built-in styles
Added the relevant sliders and drop down menu to Scrapbook style.
Because of Cg, scrapbook had to be modified to not use delayed drawing as much as possible. So we do not set the alpha to be 0.99 but maintain it at 1.0. Then the layers are not sorted by delayed drawing which is what we want.

To remove the halving of alpha for the png borders, then source file itlself was modified using the gimp.